backend/requirements.txt
fastapi
uvicorn
python-multipart
semgrep
google-generativeai
google-genai
requests
python-dotenv
# standard libs: ast, subprocess, etc.